+++ horovod/tensorflow/keras/__init__.py
@@ -14,6 +14,7 @@
 # ==============================================================================
 
 import inspect
+import warnings
 
 import tensorflow as tf
 
@@ -31,7 +32,7 @@ from horovod.tensorflow import mpi_threads_supported, mpi_enabled, mpi_built
 from horovod.tensorflow import gloo_enabled, gloo_built
 from horovod.tensorflow import nccl_built, ddl_built, ccl_built, cuda_built, rocm_built
 from horovod.tensorflow import Average, Sum
-from horovod.tensorflow import compression
+from horovod.tensorflow.compression import Compression
 
 
 import horovod._keras as _impl
@@ -49,7 +50,7 @@ except:
 
 def DistributedOptimizer(optimizer, name=None,
                          device_dense='', device_sparse='',
-                         compression=compression.Compression.none,
+                         compression=Compression.none,
                          sparse_as_dense=False,
                          gradient_predivide_factor=1.0,
                          op=Average,
@@ -149,7 +150,7 @@ def allreduce(value, name=None, average=None,
               prescale_factor=1.0,
               postscale_factor=1.0,
               op=None,
-              compression=compression.Compression.none):
+              compression=Compression.none):
     """
     Perform an allreduce on a tensor-compatible value.
 
@@ -210,7 +211,7 @@ def broadcast(value, root_rank, name=None):
     return _impl.broadcast(K, value, root_rank, name)
 
 
-def load_model(filepath, custom_optimizers=None, custom_objects=None, compression=compression.Compression.none):
+def load_model(filepath, custom_optimizers=None, custom_objects=None, compression=Compression.none):
     """
     Loads a saved Keras model with a Horovod DistributedOptimizer.